# src/main/resources/static/swagger/index.json
{
	"swagger": "2.0",
	"info": {
		"description": "openapi",
		"version": "3.0.0",
		"title": "益友会",
		"termsOfService": "https://www.yestae.com/home"
	},
	"host": "openapi-test.yestae.com",
	"basePath": "/api",
...
	"paths": {
...
		"/user/add": {
			"post": {
				"tags": ["user-api-modify-controller"],
				"summary": "新增用户",
				"description": "新增用户",
				"operationId": "addUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"name\":\"BeJson2\",\"mobile\":\"18311112223\",\"gender\":1,\"birthday\":\"2019-01-01\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
...
		"/user/updateMobile": {
			"post": {
				"tags": ["user-api-modify-controller"],
				"summary": "修改会员手机号",
				"description": "修改会员手机号",
				"operationId": "updateMobileUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"uid\":\"1128947923510472705\",\"mobile\":\"18311112226\",\"accessToken\":\"123\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/user/updatePassword": {
			"post": {
				"tags": ["user-api-modify-controller"],
				"summary": "修改会员登录密码",
				"description": "修改会员登录密码",
				"operationId": "updatePasswordUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"uid\":\"1128947923510472705\",\"newPassword\":\"qaz123\",\"accessToken\":\"2484cd8f2b541383ecd735cd577a3c9e\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"securityDefinitions": {
		"token": {
			"type": "apiKey",
			"name": "token",
			"in": "header"
		}
	}
}